LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-zh-Hant-expected.txt

FAIL navigator.language should be zh-hant. Was zh-tw.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
